+++ sktime/forecasting/confidence.py
