+++ sktime/utils/validation/tests/test_write_dataframe_to_ts.py
@@ -10,8 +10,7 @@ from pandas._testing import assert_frame_equal
 
 import sktime
 
-from sktime.utils.data_loading import load_from_tsfile_to_dataframe
-from sktime.utils.results_writing import write_dataframe_to_tsfile
+from sktime.utils.data_io import load_from_tsfile_to_dataframe, write_dataframe_to_tsfile
 
 
 def test_write_dataframe_to_ts_success(tmp_path):
